+++ src/cdk/classes/sshkey.ts
@@ -53,21 +53,21 @@ export class SSHKey implements Resource {
     }
   }
 
-  async delete(apiFactory: IAPIFactory): Promise<void> {
+  async delete(apiFactory: IAPIFactory): Promise<boolean> {
     const namespace = this.cdk?.namespace ?? "";
     const allSSHKeys = await apiFactory.sshkey.getAllSSHKeys({
       label_selector: `namespace=${namespace}`,
     });
     const sshKey = allSSHKeys.find((obj) => obj.name == this.getName());
-    if (!sshKey) return;
-    await apiFactory.sshkey.deleteSSHKey(sshKey.id);
+    if (!sshKey) return false;
+    return await apiFactory.sshkey.deleteSSHKey(sshKey.id);
   }
 
   static async deleteUnusedResources(
     localResourceNames: string[],
     namespace: string,
     apiFactory: IAPIFactory
-  ): Promise<void> {
+  ): Promise<boolean> {
     const remoteResources = await apiFactory.sshkey.getAllSSHKeys({
       label_selector: `namespace=${namespace}`,
     });
@@ -75,8 +75,10 @@ export class SSHKey implements Resource {
       (sshkey) =>
         localResourceNames.findIndex((name) => name == sshkey.name) == -1
     );
-    await Promise.all(
+    if (resourcesToBeRemoved.length == 0) return false;
+    const res = await Promise.all(
       resourcesToBeRemoved.map((obj) => apiFactory.sshkey.deleteSSHKey(obj.id))
     );
+    return res.findIndex((obj) => obj === false) != 1;
   }
 }
